refactor: separate constants
