+++ src/components/map/map.tsx
@@ -1,28 +1,17 @@
 import { useMap } from '../../hooks/useMap';
 import { useEffect, useRef } from 'react';
 import { LocationData, OfferData } from '../../types/offers';
-import { URL_MARKER_ACTIVE, URL_MARKER_DEFAULT } from '../../const';
+import { defaulCustomIcon, activeCustomIcon } from '../../const/map';
 import leaflet from 'leaflet';
 type MapProps = {
   cityLocation: LocationData;
   offers: OfferData[];
-  isHovered: string;
+  isHoveredID: string;
 };
-function Map({ cityLocation, offers, isHovered }: MapProps): JSX.Element {
+function Map({ cityLocation, offers, isHoveredID }: MapProps): JSX.Element {
   const mapRef = useRef<HTMLDivElement>(null);
 
   const map = useMap(mapRef, cityLocation);
-  const defaulCustomIcon = leaflet.icon({
-    iconUrl: URL_MARKER_DEFAULT,
-    iconSize: [27, 39],
-    iconAnchor: [20, 40],
-  });
-
-  const activeCustomIcon = leaflet.icon({
-    iconUrl: URL_MARKER_ACTIVE,
-    iconSize: [27, 39],
-    iconAnchor: [20, 40],
-  });
 
   useEffect(() => {
     if (map) {
@@ -35,13 +24,13 @@ function Map({ cityLocation, offers, isHovered }: MapProps): JSX.Element {
             },
             {
               icon:
-                offer.id === isHovered ? activeCustomIcon : defaulCustomIcon,
+                offer.id === isHoveredID ? activeCustomIcon : defaulCustomIcon,
             }
           )
           .addTo(map);
       });
     }
-  }, [map, offers, isHovered]);
+  }, [map, offers, isHoveredID]);
   return <div style={{ height: '794px', width: '500px' }} ref={mapRef}></div>;
 }
 export { Map };
