+++ src/const/auth.ts
@@ -1,0 +1,6 @@
+enum AuthorizationStatus {
+  Auth = 'AUTH',
+  NoAuth = 'NO_AUTH',
+  Unknown = 'UNKNOWN',
+}
+export { AuthorizationStatus };
