+++ src/const/routes.ts
@@ -1,0 +1,7 @@
+enum AppRoute {
+  Main = '/',
+  Login = '/login',
+  Favorites = '/favorites',
+  Offer = '/offer/',
+}
+export { AppRoute };
